LayoutTests/http/tests/websocket/tests/hybi/set-protocol.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("WebSocket.protocol should not be modified.");

window.jsTestIsAsync = true;
if (window.layoutTestController)
    layoutTestController.overridePreference("WebKitHixie76WebSocketProtocolEnabled", 0);

var ws = new WebSocket("ws://127.0.0.1:8880/websocket/tests/hybi/accept-first-subprotocol", "chat");
var closeEvent;

shouldBeEqualToString("ws.protocol", "");
ws.protocol = "superchat";
shouldBeEqualToString("ws.protocol", "");

ws.onopen = function()
{
    debug("Connected");
    shouldBeEqualToString("ws.protocol", "chat");
    ws.protocol = "superchat";
    shouldBeEqualToString("ws.protocol", "chat");
};

ws.onclose = function(event)
{
    debug("Closed");
    shouldBeEqualToString("ws.protocol", "chat");
    ws.protocol = "superchat";
    shouldBeEqualToString("ws.protocol", "chat");
    closeEvent = event;
    shouldBeTrue("closeEvent.wasClean");
    setTimeout("checkAfterOnClose()", 0);
};

function checkAfterOnClose()
{
    debug("Exited onclose handler");
    shouldBeEqualToString("ws.protocol", "chat");
    ws.protocol = "superchat";
    shouldBeEqualToString("ws.protocol", "chat");
    finishJSTest();
}

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
